move utils folder
